Address PR review feedback
- Update copyright headers on all DevCenter YAML resources to 2026.
- Move `EmitLcomValues` and `LcomTable` from top-level test classes into
  `BucketedMetricCardTest` as public static nested classes; the YAML
  pipeline references them via their `$`-qualified FQNs.
- Add `io.moderne.recipe:rewrite-prethink:latest.release` as a runtime
  `implementation` dependency so the prethink recipes referenced from
  `code-quality-devcenter.yml` (e.g. `FindClassMetrics`, `FindTestGaps`)
  resolve through this artifact's transitive closure.
- Expand the comment on `UpgradesAndMigrations.allowWritingInThisCycle`
  to document why returning `true` is safe: `BucketedMetricCard` defers
  its insert to cycle 2; all other cards default to single-cycle, and
  the existing `bestRow` dedup in `insertRow` makes any accidental
  re-run a no-op.
