+++ src/main/resources/META-INF/rewrite/examples.yml
@@ -1,4 +1,4 @@
-# Copyright 2025 the original author or authors.
+# Copyright 2026 the original author or authors.
 # <p>
 # Licensed under the Moderne Source Available License (the "License");
 # you may not use this file except in compliance with the License.
